+++ zoneinfo/zone_reader.go
@@ -5,11 +5,12 @@ package zoneinfo
 //-----------------------------------------------------------------------------
 
 type ZoneRuleReader struct {
-	f DataIO
+	f         DataIO
+	chunkSize uint8
 }
 
-func NewZoneRuleReader(f DataIO) ZoneRuleReader {
-	return ZoneRuleReader{f}
+func NewZoneRuleReader(f DataIO, chunkSize uint8) ZoneRuleReader {
+	return ZoneRuleReader{f, chunkSize}
 }
 
 func (r *ZoneRuleReader) Reset() {
@@ -17,7 +18,7 @@ func (r *ZoneRuleReader) Reset() {
 }
 
 func (r *ZoneRuleReader) Seek(i uint16) {
-	r.f.Seek(i * 11)
+	r.f.Seek(i * uint16(r.chunkSize))
 }
 
 func (r *ZoneRuleReader) Read() ZoneRuleRecord {
@@ -38,11 +39,12 @@ func (r *ZoneRuleReader) Read() ZoneRuleRecord {
 //-----------------------------------------------------------------------------
 
 type ZonePolicyReader struct {
-	f DataIO
+	f         DataIO
+	chunkSize uint8
 }
 
-func NewZonePolicyReader(f DataIO) ZonePolicyReader {
-	return ZonePolicyReader{f}
+func NewZonePolicyReader(f DataIO, chunkSize uint8) ZonePolicyReader {
+	return ZonePolicyReader{f, chunkSize}
 }
 
 func (r *ZonePolicyReader) Reset() {
@@ -50,7 +52,7 @@ func (r *ZonePolicyReader) Reset() {
 }
 
 func (r *ZonePolicyReader) Seek(i uint16) {
-	r.f.Seek(i * 4)
+	r.f.Seek(i * uint16(r.chunkSize))
 }
 
 func (r *ZonePolicyReader) Read() ZonePolicyRecord {
@@ -64,11 +66,12 @@ func (r *ZonePolicyReader) Read() ZonePolicyRecord {
 //-----------------------------------------------------------------------------
 
 type ZoneEraReader struct {
-	f DataIO
+	f         DataIO
+	chunkSize uint8
 }
 
-func NewZoneEraReader(f DataIO) ZoneEraReader {
-	return ZoneEraReader{f}
+func NewZoneEraReader(f DataIO, chunkSize uint8) ZoneEraReader {
+	return ZoneEraReader{f, chunkSize}
 }
 
 func (r *ZoneEraReader) Reset() {
@@ -76,15 +79,15 @@ func (r *ZoneEraReader) Reset() {
 }
 
 func (r *ZoneEraReader) Seek(i uint16) {
-	r.f.Seek(i * 11)
+	r.f.Seek(i * uint16(r.chunkSize))
 }
 
 func (r *ZoneEraReader) Read() ZoneEraRecord {
 	var record ZoneEraRecord
 	record.FormatIndex = r.f.ReadU16()
 	record.PolicyIndex = r.f.ReadU8()
-	record.OffsetCode = int8(r.f.ReadU8())
 	record.DeltaCode = r.f.ReadU8()
+	record.OffsetSecondsCode = int16(r.f.ReadU16())
 	record.UntilYear = int16(r.f.ReadU16())
 	record.UntilMonth = r.f.ReadU8()
 	record.UntilDay = r.f.ReadU8()
@@ -97,11 +100,12 @@ func (r *ZoneEraReader) Read() ZoneEraRecord {
 //-----------------------------------------------------------------------------
 
 type ZoneInfoReader struct {
-	f DataIO
+	f         DataIO
+	chunkSize uint8
 }
 
-func NewZoneInfoReader(f DataIO) ZoneInfoReader {
-	return ZoneInfoReader{f}
+func NewZoneInfoReader(f DataIO, chunkSize uint8) ZoneInfoReader {
+	return ZoneInfoReader{f, chunkSize}
 }
 
 func (r *ZoneInfoReader) Reset() {
@@ -109,7 +113,7 @@ func (r *ZoneInfoReader) Reset() {
 }
 
 func (r *ZoneInfoReader) Seek(i uint16) {
-	r.f.Seek(i * 12)
+	r.f.Seek(i * uint16(r.chunkSize))
 }
 
 func (r *ZoneInfoReader) Read() ZoneInfoRecord {